# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Heat Network Routing team API",
    "version": "1.0.0",
    "description": "Version 1 integration contract. Health, upload, job status, ranked variants and streamed result download are implemented. Jobs run parsing, routing, calculation, validation and export. Map remains planned."
  },
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Process is running; does not imply solver readiness.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "contractVersion",
                    "implementation"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "contractVersion": {
                      "type": "string"
                    },
                    "implementation": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "post": {
        "operationId": "createJob",
        "x-implementation": "implemented",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "2d",
                      "depth"
                    ],
                    "default": "2d"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Upload stored; asynchronous processing queued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Relative URL of this job status",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "File or entire multipart request exceeds the configured byte limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Mode unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Queue capacity reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds before retrying",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Server storage or processing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Request must use multipart/form-data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Accepts the upload, then asynchronously validates, routes, calculates and exports. SUCCEEDED means the result and summaries have been saved. Partial connection is reported in summaries and diagnostics."
      }
    },
    "/api/jobs/{jobId}": {
      "parameters": [
        {
          "name": "jobId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "getJob",
        "x-implementation": "implemented",
        "responses": {
          "200": {
            "description": "Current task state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "description": "no-store: poll for fresh status",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Unknown job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server storage or processing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{jobId}/variants": {
      "parameters": [
        {
          "name": "jobId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "listVariants",
        "x-implementation": "implemented",
        "responses": {
          "200": {
            "description": "Ranked summaries using official snake_case fields.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "maxItems": 3,
                  "items": {
                    "$ref": "#/components/schemas/VariantSummary"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Result not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Storage or internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{jobId}/result": {
      "parameters": [
        {
          "name": "jobId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "downloadResult",
        "x-implementation": "implemented",
        "responses": {
          "200": {
            "description": "Streamed GeoJSON attachment, up to 500 MiB.",
            "content": {
              "application/geo+json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Unknown job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Result not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Storage or internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{jobId}/map": {
      "parameters": [
        {
          "name": "jobId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "getMapPage",
        "x-implementation": "planned",
        "parameters": [
          {
            "name": "layer",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "input",
                "result"
              ]
            }
          },
          {
            "name": "variantId",
            "in": "query",
            "description": "Required for result layer; server-issued variant IDs are strings.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": true,
            "description": "minLon,minLat,maxLon,maxLat in EPSG:4326",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5000,
              "default": 1000
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bounded page; follow nextCursor with identical query until null.",
            "content": {
              "application/geo+json": {
                "schema": {
                  "$ref": "#/components/schemas/MapPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid map query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown job or variant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Data not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Includes input diagnostic codes, FILE_TOO_LARGE, UNSUPPORTED_MODE, QUEUE_FULL, JOB_NOT_FOUND, RESULT_NOT_READY, ROUTE_NOT_FOUND, OUTPUT_LIMIT_EXCEEDED, RESULT_UNAVAILABLE, JOB_INTERRUPTED, SERVER_RESTARTED and INTERNAL_ERROR. Legacy PROCESSING_UNAVAILABLE may remain in stored jobs from older versions. This is an extensible string, not a closed enum."
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Optional diagnostic references. inputObjectId preserves the original JSON number/string type; candidateId and segmentId refer to internal candidate/segment IDs. No filesystem paths or stack traces are returned."
          }
        }
      },
      "MapPage": {
        "type": "object",
        "required": [
          "type",
          "features",
          "nextCursor"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "FeatureCollection"
            ]
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MapFeature"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "VariantSummary": {
        "type": "object",
        "required": [
          "id",
          "object_type",
          "variant_id",
          "rank",
          "construction_cost",
          "chamber_construction_cost",
          "existing_chamber_tie_in_count",
          "existing_chamber_tie_in_cost",
          "unconnected_penalty",
          "calculated_cost",
          "new_network_length",
          "score",
          "unconnected_oks_ids"
        ],
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "object_type": {
            "type": "string",
            "enum": [
              "variant_summary"
            ]
          },
          "variant_id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "rank": {
            "type": "integer"
          },
          "construction_cost": {
            "type": "number"
          },
          "chamber_construction_cost": {
            "type": "number"
          },
          "existing_chamber_tie_in_count": {
            "type": "integer"
          },
          "existing_chamber_tie_in_cost": {
            "type": "number"
          },
          "unconnected_penalty": {
            "type": "number"
          },
          "calculated_cost": {
            "type": "number"
          },
          "new_network_length": {
            "type": "number"
          },
          "score": {
            "type": "number"
          },
          "unconnected_oks_ids": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            }
          }
        }
      },
      "Geometry": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "type",
              "coordinates"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Point"
                ]
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "number"
                },
                "minItems": 2,
                "maxItems": 2
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type",
              "coordinates"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "LineString"
                ]
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": {
                    "type": "number"
                  },
                  "minItems": 2,
                  "maxItems": 2
                },
                "minItems": 2
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type",
              "coordinates"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "MultiLineString"
                ]
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "minItems": 2,
                    "maxItems": 2
                  },
                  "minItems": 2
                },
                "minItems": 1
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type",
              "coordinates"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Polygon"
                ]
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "minItems": 2,
                    "maxItems": 2
                  },
                  "minItems": 4
                },
                "minItems": 1
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type",
              "coordinates"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "MultiPolygon"
                ]
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "number"
                      },
                      "minItems": 2,
                      "maxItems": 2
                    },
                    "minItems": 4
                  },
                  "minItems": 1
                },
                "minItems": 1
              }
            }
          }
        ]
      },
      "MapFeature": {
        "type": "object",
        "required": [
          "type",
          "geometry",
          "properties"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "Feature"
            ]
          },
          "geometry": {
            "$ref": "#/components/schemas/Geometry"
          },
          "properties": {
            "type": "object",
            "required": [
              "id",
              "object_type"
            ],
            "properties": {
              "id": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  }
                ]
              },
              "object_type": {
                "type": "string",
                "enum": [
                  "source",
                  "heat_network",
                  "heat_chamber",
                  "oks_connection_point",
                  "restriction",
                  "technical_node"
                ]
              },
              "variant_id": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  }
                ]
              },
              "diameter": {
                "type": "integer"
              },
              "flow_tph": {
                "type": "number"
              },
              "restriction_type": {
                "type": "string"
              },
              "cost": {
                "type": "number"
              },
              "length": {
                "type": "number"
              },
              "laying_method": {
                "type": "string",
                "enum": [
                  "base",
                  "special"
                ]
              }
            },
            "additionalProperties": true
          }
        }
      },
      "Job": {
        "type": "object",
        "required": [
          "jobId",
          "status",
          "stage",
          "mode",
          "diagnostics"
        ],
        "properties": {
          "jobId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "stage": {
            "type": "string",
            "enum": [
              "QUEUED",
              "VALIDATING",
              "ROUTING",
              "CALCULATING",
              "EXPORTING",
              "DONE",
              "FAILED"
            ]
          },
          "mode": {
            "type": "string",
            "enum": [
              "2d",
              "depth"
            ]
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "QUEUED is returned after the upload is stored. VALIDATING parses the input asynchronously. SUCCEEDED/DONE is reserved for finished calculation and export; validation alone cannot succeed a job. Terminal states are retained for 24 hours by default; expired or unknown jobs return 404."
      }
    }
  }
}